📌(pysaml2) pin pysaml2 to avoid double-signing of AuthNRequests
The current version of pysaml2 (7.5.2) has an issue where AuthNRequests
are both signed in the XML and with an extra `Signature` queryparam.
This was reported initially in 2021:
    https://github.com/IdentityPython/pysaml2/issues/819
And it was fixed by a changed in SATOSA:
    https://github.com/IdentityPython/SATOSA/pull/380
But it reappeared apparently and the original reporter has a PR open
against pysaml2 that is supposed to fix it:
    https://github.com/IdentityPython/pysaml2/pull/973
They report that the regression was introduced in pysaml2 by
   https://github.com/IdentityPython/pysaml2/pull/834
We try here to pin pysaml2 to the last version before this PR was
merged. Unfortunately this is quite an old version, but from basic
testing it seems to still be compatible with the current SATOSA
version.
This in turn forces us to also pin xmlschema to avoid
    https://github.com/IdentityPython/pysaml2/issues/947

Hopefully this can be temporary.